+++ src/Migration/Migration1749821607InitialSetup.php
@@ -17,21 +17,20 @@ class Migration1749821607InitialSetup extends MigrationStep
         $schemaManager = $connection->createSchemaManager();
         $columns = array_map(fn($col) => $col->getName(), $schemaManager->listTableColumns('product'));
 
-        if (!in_array('flip', $columns, true) || !in_array('flip_id', $columns, true)) {
+        if (!in_array('flip_id', $columns, true)) {
             $connection->executeStatement('
-            ALTER TABLE `product`
-            ADD COLUMN IF NOT EXISTS `flip` BINARY(16),
-            ADD COLUMN IF NOT EXISTS `flip_id` BINARY(16) NULL
-        ');
+                ALTER TABLE `product`
+                ADD COLUMN `flip_id` BINARY(16) NULL
+            ');
         }
 
         $foreignKeys = array_map(fn($fk) => $fk->getName(), $schemaManager->listTableForeignKeys('product'));
         if (!in_array('fk.product.flip_id', $foreignKeys, true)) {
             $connection->executeStatement('
-            ALTER TABLE `product`
-            ADD CONSTRAINT `fk.product.flip_id` FOREIGN KEY (`flip_id`)
-            REFERENCES `product_media` (`id`) ON DELETE SET NULL ON UPDATE CASCADE
-        ');
+                ALTER TABLE `product`
+                ADD CONSTRAINT `fk.product.flip_id` FOREIGN KEY (`flip_id`)
+                REFERENCES `product_media` (`id`) ON DELETE SET NULL ON UPDATE CASCADE
+            ');
         }
     }
 }