+++ src/Resources/app/administration/src/module/sw-product-image/index.js
@@ -7,11 +7,22 @@ Shopware.Component.override('sw-product-image', {
     compatConfig: Shopware.compatConfig,
 
     emits: [
+        'sw-product-image-cover',
+        'sw-product-image-delete',
         'sw-product-image-flip',
+        'sw-product-image-flip-delete',
+        'sw-product-image-context-photo',
+        'sw-product-image-context-photo-delete',
     ],
 
     props: {
         isFlip: {
+            type: Boolean,
+            required: false,
+            default: false,
+        },
+
+        isContextPhoto: {
             type: Boolean,
             required: false,
             default: false,